requirements.txt
python-dotenv
boto3
sentence-transformers
pillow
pymilvus
